chore: move interfaces to utils
